moved knex connection middleware into utils
